modularized script, now support ssd mobilenet v1/v2 retrained on any version of object detection api, added missing docstring, download model on script automatically
