--- training/tensorrt_generation/download_model.sh
@@ -1,1 +1,0 @@
-wget https://raw.githubusercontent.com/neuralet/neuralet-models/master/amd64/ped_ssd_mobilenet_v2/frozen_inference_graph.pb -O ssd_mobilenet_v2_pedestrian.pb
